feat: implement guest management APIs for bookmarks, comments, and highlights
