+++ apps/web/src/routes/kiwi/-reader/apis/bookmark.ts
@@ -2,16 +2,19 @@ import { toast } from "sonner";
 
 import { Bookmark } from "@bookiwi/supabase/types";
 
+import {
+  addGuestBookmark,
+  getGuestBookmarks,
+  removeGuestBookmark,
+} from "./guest";
+
 import supabaseManager from "#/managers/supabase";
 import userManager from "#/managers/user";
 
 export const addBookmark = async (newBookmark: Bookmark) => {
   try {
     if (userManager.isGuest) {
-      userManager.setGuestBookmarks([
-        ...userManager.getGuestBookmarks(),
-        newBookmark,
-      ]);
+      addGuestBookmark(newBookmark);
       return;
     }
     await supabaseManager.reader.addBookmark(newBookmark);
@@ -31,11 +34,7 @@ export const removeBookmark = async ({
 }) => {
   try {
     if (userManager.isGuest) {
-      userManager.setGuestBookmarks(
-        userManager
-          .getGuestBookmarks()
-          .filter((b) => b.cfiStart !== cfiStart || b.cfiEnd !== cfiEnd),
-      );
+      removeGuestBookmark({ cfiStart, cfiEnd });
       return;
     }
     await supabaseManager.reader.removeBookmark({
@@ -51,10 +50,11 @@ export const removeBookmark = async ({
 export const getBookmarks = async (participantId: string) => {
   try {
     if (userManager.isGuest) {
-      const userBookmarks = userManager.getGuestBookmarks();
-      return userBookmarks;
+      const bookmarks = getGuestBookmarks();
+      return bookmarks;
     }
-    return await supabaseManager.reader.getBookmarks(participantId);
+    const bookmarks = await supabaseManager.reader.getBookmarks(participantId);
+    return bookmarks;
   } catch (error) {
     toast.error("북마크 정보를 불러오는데 실패했습니다.");
     return [];
